PerfectUniversity/ConTip-master/requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-filters==0.2.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.5
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wrapt==1.11.2
